#  Optimization: Devirtualize Drawable attributes - Drawables 1 of N