--- a/bench/draw.cpp
+++ b/bench/draw.cpp
@@ -1,0 +1,49 @@
+#include <benchmark/benchmark.h>
+#include <font.h>
+#include <rect.h>
+#include <bitmap.h>
+#include <sprite.h>
+#include <graphics.h>
+
+constexpr int num_sprites = 5000;
+
+class TestSprite : public Drawable {
+	public:
+		void Draw() override {}
+		int GetZ() const override { return 0; }
+		DrawableType GetType() const override { return TypeDefault; }
+		bool IsGlobal() const override { return true; }
+};
+
+static void BM_DrawSort(benchmark::State& state) {
+	std::vector<std::unique_ptr<TestSprite>> sprites;
+	for (int i = 0; i < num_sprites; ++i) {
+		sprites.push_back(std::make_unique<TestSprite>());
+	}
+	Graphics::DrawableList list;
+	for (auto& s: sprites) {
+		list.push_back(s.get());
+	}
+
+	for (auto _: state) {
+		Graphics::SortDrawableList(list);
+	}
+}
+
+BENCHMARK(BM_DrawSort);
+
+static void BM_DrawSortLocality(benchmark::State& state) {
+	std::array<TestSprite,num_sprites> sprites;
+	Graphics::DrawableList list;
+	for (auto& s: sprites) {
+		list.push_back(&s);
+	}
+
+	for (auto _: state) {
+		Graphics::SortDrawableList(list);
+	}
+}
+
+BENCHMARK(BM_DrawSortLocality);
+
+BENCHMARK_MAIN();