--- a/src/background.cpp
+++ b/src/background.cpp
@@ -29,6 +29,7 @@
 #include "output.h"
 
 Background::Background(const std::string& name) :
+	Drawable(TypeBackground, Priority_Background, false),
 	visible(true), tone_effect(Tone()),
 	bg_hscroll(0), bg_vscroll(0), bg_x(0), bg_y(0),
 	fg_hscroll(0), fg_vscroll(0), fg_x(0), fg_y(0) {
@@ -44,6 +45,7 @@ Background::Background(const std::string& name) :
 }
 
 Background::Background(int terrain_id) :
+	Drawable(TypeBackground, Priority_Background, false),
 	visible(true), tone_effect(Tone()),
 	bg_hscroll(0), bg_vscroll(0), bg_x(0), bg_y(0),
 	fg_hscroll(0), fg_vscroll(0), fg_x(0), fg_y(0) {
@@ -103,14 +105,6 @@ void Background::OnForegroundFrameGraphicReady(FileRequestResult* result) {
 
 Background::~Background() {
 	Graphics::RemoveDrawable(this);
-}
-
-int Background::GetZ() const {
-	return z;
-}
-
-DrawableType Background::GetType() const {
-	return type;
 }
 
 Tone Background::GetTone() const {