--- a/src/background.h
+++ b/src/background.h
@@ -36,13 +36,7 @@ public:
 	Tone GetTone() const;
 	void SetTone(Tone tone);
 
-	int GetZ() const override;
-	DrawableType GetType() const override;
-
 private:
-	static const int z = Priority_Background;
-	static const DrawableType type = TypeBackground;
-
 	static void Update(int& rate, int& value);
 	static int Scale(int x);