--- a/src/battle_animation.h
+++ b/src/battle_animation.h
@@ -90,7 +90,6 @@ protected:
 class BattleAnimationMap : public BattleAnimation {
 public:
 	BattleAnimationMap(const RPG::Animation& anim, Game_Character& target, bool global);
-	~BattleAnimationMap() override;
 	void Draw() override;
 protected:
 	void FlashTargets(int r, int g, int b, int p) override;
@@ -106,7 +105,6 @@ protected:
 class BattleAnimationBattle : public BattleAnimation {
 public:
 	BattleAnimationBattle(const RPG::Animation& anim, std::vector<Game_Battler*> battlers, bool only_sound = false, int cutoff_frame = -1);
-	~BattleAnimationBattle() override;
 	void Draw() override;
 protected:
 	void FlashTargets(int r, int g, int b, int p) override;