--- a/src/drawable.cpp
+++ b/src/drawable.cpp
@@ -17,6 +17,19 @@
 
 #include "drawable.h"
 #include "rpg_savepicture.h"
+#include "graphics.h"
+
+Drawable::Drawable(DrawableType type, int z, bool is_global)
+	: _z(z),
+	_type(static_cast<decltype(_type)>(type)),
+	_is_global(is_global)
+{
+}
+
+void Drawable::SetZ(int nz) {
+	if (_z != nz) Graphics::UpdateZCallback();
+	_z = nz;
+}
 
 int Drawable::GetPriorityForMapLayer(int which) {
 	switch (which) {