--- a/src/fps_overlay.cpp
+++ b/src/fps_overlay.cpp
@@ -30,10 +30,6 @@ FpsOverlay::FpsOverlay() :
 	Graphics::RegisterDrawable(this);
 }
 
-FpsOverlay::~FpsOverlay() {
-	Graphics::RemoveDrawable(this);
-}
-
 void FpsOverlay::Update() {
 	int mod = Player::GetSpeedModifier();