--- a/src/frame.cpp
+++ b/src/frame.cpp
@@ -25,7 +25,9 @@
 #include "main_data.h"
 #include "frame.h"
 
-Frame::Frame() {
+Frame::Frame() :
+	Drawable(TypeFrame, Priority_Frame, false)
+{
 	if (!Data::system.frame_name.empty() && Data::system.show_frame) {
 		FileRequestAsync* request = AsyncHandler::RequestFile("Frame", Data::system.frame_name);
 		request->SetGraphicFile(true);
@@ -38,14 +40,6 @@ Frame::Frame() {
 
 Frame::~Frame() {
 	Graphics::RemoveDrawable(this);
-}
-
-int Frame::GetZ() const {
-	return z;
-}
-
-DrawableType Frame::GetType() const {
-	return type;
 }
 
 void Frame::Update() {