--- a/src/message_overlay.h
+++ b/src/message_overlay.h
@@ -46,12 +46,6 @@ public:
 
 	void Draw() override;
 
-	int GetZ() const override;
-
-	DrawableType GetType() const override;
-
-	bool IsGlobal() const override;
-
 	void Update();
 
 	void AddMessage(const std::string& message, Color color);
@@ -61,12 +55,9 @@ public:
 private:
 	bool IsAnyMessageVisible() const;
 
-	DrawableType type;
-
 	BitmapRef bitmap;
 	BitmapRef black;
 
-	int z;
 	int ox;
 	int oy;