--- a/src/plane.cpp
+++ b/src/plane.cpp
@@ -24,9 +24,8 @@
 #include "game_map.h"
 
 Plane::Plane() :
-	type(TypePlane),
+	Drawable(TypePlane, 0, false),
 	visible(true),
-	z(0),
 	ox(0),
 	oy(0) {
 
@@ -111,13 +110,6 @@ bool Plane::GetVisible() const {
 void Plane::SetVisible(bool nvisible) {
 	visible = nvisible;
 }
-int Plane::GetZ() const {
-	return z;
-}
-void Plane::SetZ(int nz) {
-	if (z != nz) Graphics::UpdateZCallback();
-	z = nz;
-}
 int Plane::GetOx() const {
 	return ox;
 }
@@ -142,6 +134,3 @@ void Plane::SetTone(Tone tone) {
 	}
 }
 
-DrawableType Plane::GetType() const {
-	return type;
-}