This devirtualizes `Drawable::GetZ()` resulting in 3x speed up when we sort the drawables.

Benchmark results on a Surface Go Tablet running Windows 10 / WSL.

Before:
```
---------------------------------------------------
Benchmark            Time           CPU Iterations
---------------------------------------------------
BM_DrawSort     320284 ns     313895 ns       2240
```

After:
```
---------------------------------------------------
Benchmark            Time           CPU Iterations
---------------------------------------------------
BM_DrawSort     103358 ns     104627 ns       7467
```

And for fun, if we keep `GetZ()` as non-virtual but don't inline it:
```
---------------------------------------------------
Benchmark            Time           CPU Iterations
---------------------------------------------------
BM_DrawSort     269318 ns     272770 ns       2635
```