@fmatthew5876 2019-12-07
On my linux desktop

before:
```
---------------------------------------------------
Benchmark            Time           CPU Iterations
---------------------------------------------------
BM_DrawSort     120860 ns     120858 ns       5877
```

After:
```
---------------------------------------------------
Benchmark            Time           CPU Iterations
---------------------------------------------------
BM_DrawSort      43942 ns      43939 ns      13191
```